requirements.txt
mysql_connector_repackaged==0.3.1
twitchio==2.10.0
